Closes #51
Some NPI records had multiple tibbles in the taxonomies list column of the results with FALSE for all values of the `primary` column. Hence, there was no record with a primary taxonomy. The fix here was to return the first row of the taxonomies tibble grouped by NPI. This removes the error by ensuring a 1-to-1 relationship between NPIs and taxonomy rows.